+++ centrifuge/structure/file.py
@@ -9,7 +9,7 @@ from centrifuge.structure import BaseStorage
 from tornado.options import define
 
 define(
-    "path", default='structure.json', help="Path to JSON file with structure configuration", type=str
+    "file", default='structure.json', help="Path to JSON file with structure configuration", type=str
 )
 
 
@@ -19,7 +19,7 @@ class Storage(BaseStorage):
 
     def __init__(self, *args, **kwargs):
         super(Storage, self).__init__(*args, **kwargs)
-        self.data = json.load(open(self.options.path, 'r'))
+        self.data = json.load(open(self.options.file, 'r'))
 
     def connect(self, callback=None):
         callback()