+++ centrifuge/__init__.py
@@ -1,4 +1,4 @@
 # coding: utf-8
 # Copyright (c) Alexandr Emelin. MIT license.
 
-__version__ = "0.7.0dev"
+__version__ = "0.7.0"